Fixes for LINK, maximum gold value in a transaction, house system-related exceptions (Issue #552). - Fixed: LINK.* not returning error if LINK was not set or invalid. - Fixed: Multi house sign's LINK wasn't set on multi creation.

+++ src/game/chars/CChar.h
@@ -1213,7 +1213,7 @@ public:
 	bool NPC_OnTrainPay( CChar * pCharSrc, CItemMemory * pMemory, CItem * pGold );
 	bool NPC_OnTrainHear( CChar * pCharSrc, lpctstr pCmd );
 	bool NPC_TrainSkill( CChar * pCharSrc, SKILL_TYPE skill, ushort uiAmountToTrain );
-    int PayGold(CChar * pCharSrc, int iGold, CItem * pGold, ePayGold iReason);
+    int64 PayGold(CChar * pCharSrc, int64 iGold, CItem * pGold, ePayGold iReason);
 private:
 	bool NPC_CheckWalkHere( const CPointMap & pt, const CRegion * pArea ) const;
 	void NPC_OnNoticeSnoop( const CChar * pCharThief, const CChar * pCharMark );
